requirements.txt
Django==4.1.2
serializers==0.2.4
djangorestframework==3.14.0